\name{nordklimdata1-package}
\alias{nordklimdata1-package}
\alias{nordklimdata1}
\docType{package}
\title{
Nordklim data set 1.0
}
\description{
The NORDKLIM data set contains close to 70 000 years of monthly data from 114 stations. The
station network covers all five Nordic countries, including data from the Faeroe Islands, Jan Mayen,
Bjornoya and Svalbard. There are seven monthly climatic elements describing temperature,
two on precipitation and one on air pressure, cloud cover and snow cover.

AUTHORS:
Heikki Tuomenvirta, Achim Drebs, Eirik Forland, Ole Einar Tveito, Hans
Alexandersson, Ellen Vaarby Laursen and Trausti Jo'nsson.

PROJECT CONTRACTORS:
NORDKLIM/NORDMET on behalf of the National meteorological services in
Denmark (DMI), Finland (FMI), Iceland (VI), Norway (DNMI) and Sweden (SMHI)
}
\details{
\tabular{ll}{
Package: \tab nordklimdata1\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-10\cr
License: \tab GPL (>= 3)\cr
}
}
\author{
Author: Jose' Gama

Maintainer: jgama@abo.fi
}
\source{\url{http://www.smhi.se/hfa_coord/nordklim}}
\references{
Nordklim dataset 1.0 - description and illustrations
Norwegian meteorological institute, 08/01 KLIMA, 2001
}
\keyword{ package }

